Creating VisitInvites styles file
