+++ src/Styles/VisitInvistesStyles.js
@@ -1,0 +1,82 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+    principal: {
+      flex: 1,
+    },
+  
+    content: {
+      flex: 1,
+      paddingTop: 6,
+      backgroundColor: 'white',
+    },
+  
+    listSchedule: {
+      flex: 1,
+      marginHorizontal: 15,
+      marginVertical: 10,
+      borderColor: 'black',
+      borderWidth: 1,
+      borderRadius: 3,
+      backgroundColor: '#FAFAFA',
+      justifyContent: 'space-between',
+      flexDirection: 'row',
+      alignItems: 'center',
+    },
+  
+    dialogButtonStyle: {
+      marginVertical: -10,
+    },
+  
+    textBox: {
+      flex: 4,
+      paddingLeft: 4,
+      justifyContent: 'flex-start',
+      marginRight: 15,
+    },
+  
+    text: {
+      fontSize: 15,
+      paddingVertical: 2,
+    },
+  
+    buttonBox: {
+      borderColor: 'black',
+      borderWidth: 0.8,
+      borderRadius: 7,
+      backgroundColor: '#4cd964',
+      padding: 8,
+      marginRight: 15,
+      justifyContent: 'center',
+      marginVertical: 5,
+    },
+  
+    buttonInvitees: {
+      borderColor: 'black',
+      borderWidth: 0.8,
+      borderRadius: 7,
+      backgroundColor: 'white',
+      padding: 8,
+      justifyContent: 'center',
+      marginRight: 15,
+      marginVertical: 5,
+    },
+  
+    buttonText: {
+      fontSize: 12,
+      textAlign: 'center',
+    },
+  
+    footerPopUp: {
+      backgroundColor: '#F9F9FB',
+      borderColor: '#DAD9DC',
+      borderTopWidth: 0.5,
+      borderBottomLeftRadius: 8,
+      borderBottomRightRadius: 8,
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+});
+
+export default styles;
