+++ src/screens/VisitInvites.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import axios from 'axios';
-import { View, ActivityIndicator, ScrollView, StyleSheet, BackHandler } from 'react-native';
+import { View, ActivityIndicator, ScrollView, BackHandler } from 'react-native';
 import PopupDialog, { DialogTitle, DialogButton } from 'react-native-popup-dialog';
 import VisitCard from '../components/VisitCard';
 import Header from '../components/Header';
@@ -10,87 +10,9 @@ import { SCHOOL_ENDPOINT, UNABLE_TO_FIND_SCHOOL } from '../constants/generalCons
 import { logInfo, logWarn } from '../../logConfig/loggers';
 import ShowToast from '../components/Toast';
 import { backHandlerPop } from '../NavigationFunctions';
+import styles from '../Styles/VisitInvistesStyles'
 
 const FILE_NAME = 'VisitInvites';
-
-const styles = StyleSheet.create({
-  principal: {
-    flex: 1,
-  },
-
-  content: {
-    flex: 1,
-    paddingTop: 6,
-    backgroundColor: 'white',
-  },
-
-  listSchedule: {
-    flex: 1,
-    marginHorizontal: 15,
-    marginVertical: 10,
-    borderColor: 'black',
-    borderWidth: 1,
-    borderRadius: 3,
-    backgroundColor: '#FAFAFA',
-    justifyContent: 'space-between',
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-
-  dialogButtonStyle: {
-    marginVertical: -10,
-  },
-
-  textBox: {
-    flex: 4,
-    paddingLeft: 4,
-    justifyContent: 'flex-start',
-    marginRight: 15,
-  },
-
-  text: {
-    fontSize: 15,
-    paddingVertical: 2,
-  },
-
-  buttonBox: {
-    borderColor: 'black',
-    borderWidth: 0.8,
-    borderRadius: 7,
-    backgroundColor: '#4cd964',
-    padding: 8,
-    marginRight: 15,
-    justifyContent: 'center',
-    marginVertical: 5,
-  },
-
-  buttonInvitees: {
-    borderColor: 'black',
-    borderWidth: 0.8,
-    borderRadius: 7,
-    backgroundColor: 'white',
-    padding: 8,
-    justifyContent: 'center',
-    marginRight: 15,
-    marginVertical: 5,
-  },
-
-  buttonText: {
-    fontSize: 12,
-    textAlign: 'center',
-  },
-
-  footerPopUp: {
-    backgroundColor: '#F9F9FB',
-    borderColor: '#DAD9DC',
-    borderTopWidth: 0.5,
-    borderBottomLeftRadius: 8,
-    borderBottomRightRadius: 8,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
 
 class VisitInvites extends React.Component {
   constructor(props) {
